Use `--dump-data` instead of an auto-generated mod.
The `--dump-data` flag was added to Factorio in version
[1.1.77](https://wiki.factorio.com/Version_history/1.1.0#1.1.77). This
simplifies the whole approach significantly and makes it almost
unnecessary to use the `factorio_exporter` at all. There is still the
benefit of managing installed mods programmatically. We also still
support conversion of the output to YAML, which is marginally useful.

Another flag `--dump-icon-sprites` was added as well that dumps
pre-rendered icon images. This isn't used, yet, but we could use it to
allow exporting a ZIP of icons.

Fixes #89 (sort of)